--- src/components/Footer/Footer.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function Footer() {
-  return (
-    <div>
-      footer
-    </div>
-  )
-}
