+++ src/components/Products/Products.js
@@ -1,0 +1,54 @@
+import React from "react";
+import useProducts from "../../model/Products";
+import { Link } from "react-router-dom";
+
+const Products = () => {
+  const { activeProducts, ...productService } = useProducts();
+  const onProductAdd = () => {
+    productService.addRandomProduct();
+  };
+
+  const onSelectedChange = (p) => {
+    productService.selectProduct(p);
+  };
+
+  const onProductFilterByName = () => {
+    productService.fliterProductByName();
+  };
+
+  const onProductAll = () => {
+    productService.getAll();
+  };
+
+  const onProductsDelete = () => {
+    productService.deleteSelected();
+  };
+
+  return (
+    <div className="App">
+      <div>
+        {activeProducts.map((p) => (
+          <div key={p.id}>
+            <input
+              type="checkbox"
+              className="chk"
+              checked={p.selected}
+              onChange={() => onSelectedChange(p)}
+            ></input>
+            <label className="product">
+              <Link to={`/products/${p.id}`}>
+                {p.name} {p.id}
+              </Link>
+            </label>
+          </div>
+        ))}
+      </div>
+      <button onClick={onProductAdd}>Add</button>
+      <button onClick={onProductFilterByName}>Filter By Kang</button>
+      <button onClick={onProductAll}>Get all</button>
+      <button onClick={onProductsDelete}>Delete By Checkbox</button>
+    </div>
+  );
+};
+
+export default Products;
